Downloads/pulse 앱/pulse/backend/requirements.txt
# FastAPI Core
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.9
pydantic-settings==2.10.1
email-validator==2.2.0

# Database
sqlalchemy==2.0.41
alembic==1.16.4
psycopg[binary]==3.2.3

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.20

# CORS
python-dotenv==1.1.1

# Development
pytest==8.4.2
httpx==0.28.1
